+++ weave/scorers/__init__.py
@@ -9,7 +9,7 @@ from weave.scorers.classification_scorer import (
 from weave.scorers.guardrails import PromptInjectionLLMGuardrail
 from weave.scorers.hallucination_scorer import HallucinationFreeScorer
 from weave.scorers.json_scorer import ValidJSONScorer
-from weave.scorers.llm_scorer import (
+from weave.scorers.scorer_types import (
     LLMScorer,
 )
 from weave.scorers.moderation_scorer import (
