+++ weave/scorers/guardrails/prompt_injection_guardrail.py
@@ -9,7 +9,7 @@ from weave.scorers.guardrails.prompts import (
     PROMPT_INJECTION_GUARDRAIL_SYSTEM_PROMPT,
     PROMPT_INJECTION_SURVEY_PAPER_SUMMARY,
 )
-from weave.scorers.llm_scorer import LLMScorer
+from weave.scorers.scorer_types import LLMScorer
 from weave.scorers.utils import stringify
