+++ weave/scorers/hallucination_scorer.py
@@ -3,7 +3,7 @@ from pydantic import BaseModel, Field
 
 import weave
 from weave.scorers.default_models import OPENAI_DEFAULT_MODEL
-from weave.scorers.llm_scorer import LLMScorer
+from weave.scorers.scorer_types import LLMScorer
 from weave.scorers.utils import stringify
 
 DEFAULT_HALLUCINATION_SYSTEM_PROMPT = """
