+++ weave/scorers/ragas_scorer.py
@@ -7,7 +7,7 @@ from pydantic import BaseModel, Field
 
 import weave
 from weave.scorers.default_models import OPENAI_DEFAULT_MODEL
-from weave.scorers.llm_scorer import LLMScorer
+from weave.scorers.scorer_types import LLMScorer
 
 
 class EntityExtractionResponse(BaseModel):
