+++ weave/scorers/scorer_types.py
@@ -75,7 +75,7 @@ class HuggingFacePipelineScorer(weave.Scorer):
     )
     model_name_or_path: str = Field(default="", description="The path to the model")
     device: Union[str, Any] = Field(
-        default="auto", description="The device to use for the model"
+        default="auto", description="The device to use for the model", validate_default=True
     )
     _pipeline: Optional["Pipeline"] = PrivateAttr(default=None)
 
@@ -107,7 +107,7 @@ class HuggingFaceScorer(weave.Scorer):
 
     model_name_or_path: str = Field(default="", description="The path to the model")
     device: Union[str, Any] = Field(
-        default="auto", description="The device to use for the model"
+        default="auto", description="The device to use for the model", validate_default=True
     )
     _model: Optional["PreTrainedModel"] = PrivateAttr(default=None)
     _tokenizer: Optional["PreTrainedTokenizer"] = PrivateAttr(default=None)
